Add XMPPSplice build flag, for horrible windows hack
This should allow building with the XMPP UI on windows to get TH splices,
without actually linking in the XMPP libraries, which currently fails
due to <https://ghc.haskell.org/trac/ghc/ticket/8830>

Then, using those splices, the EvilSplicer can make a source tree that can
be built, with XMPP libraries, bypassing that bug.

Too many ifdefs and stub functions to be happy with this. It's a massively
grody workaround.

This commit was sponsored by an anonymous bitcoin doner.
Yeah, I'd not want to have my name of this commit either... ;)