unreliable test that storeKey is atomic
The best I could come up with was a test that stored a
key and had another thread repeatedly check if the object was present on
the remote, and log the results and timestamps. It then becomes a
statistical problem, somewhere toward the end of the log it's ok if the key
has become present .. but too early might indicate that it wasn't stored
atomically. Perhaps it's my poor knowledge of statistics, but I could not
find a solution that analized the log and reliably detected non-atomic storage.